+++ buildSrc/src/main/kotlin/Config.kt
@@ -74,7 +74,7 @@ object Versions {
     const val cmake = "3.18.1" // Core requires minimum 3.15, but 3.18.1 is available through the Android SDK
     const val coroutines = "1.5.2-native-mt" // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core
     const val detektPlugin = "1.17.1" // https://github.com/detekt/detekt
-    const val dokka = "1.4.32" // https://github.com/Kotlin/dokka
+    const val dokka = "1.5.30" // https://github.com/Kotlin/dokka
     const val gradlePluginPublishPlugin = "0.15.0" // https://plugins.gradle.org/plugin/com.gradle.plugin-publish
     const val junit = "4.13.2" // https://mvnrepository.com/artifact/junit/junit
     const val jvmTarget = "1.8"