+++ packages/cinterop/src/commonMain/kotlin/io/realm/mongodb/AppException.kt
@@ -18,4 +18,4 @@ package io.realm.mongodb
 
 @Suppress("ForbiddenComment")
 // TODO: implement actual AppException with sync error codes hierarchy
-class AppException : Exception()
+class AppException(message: String?) : Exception(message)