+++ packages/cinterop/src/jvm/jni/java_class_global_def.hpp
@@ -46,8 +46,8 @@ private:
         , m_io_realm_network_transport(env, "io/realm/internal/interop/sync/NetworkTransport", false)
         , m_io_realm_response(env, "io/realm/internal/interop/sync/Response", false)
         , m_io_realm_long_pointer_wrapper(env, "io/realm/internal/interop/LongPointerWrapper", false)
-        , m_io_realm_app_exception(env, "io/realm/mongodb/AppException", false)
         , m_io_realm_sync_exception(env, "io/realm/mongodb/SyncException", false)
+        , m_io_realm_mongodb_app_exception(env, "io/realm/mongodb/AppException", false)
         , m_io_realm_sync_log_callback(env, "io/realm/internal/interop/SyncLogCallback", false)
         , m_io_realm_sync_error_callback(env, "io/realm/internal/interop/SyncErrorCallback", false)
     {
@@ -57,8 +57,8 @@ private:
     jni_util::JavaClass m_io_realm_network_transport;
     jni_util::JavaClass m_io_realm_response;
     jni_util::JavaClass m_io_realm_long_pointer_wrapper;
-    jni_util::JavaClass m_io_realm_app_exception;
     jni_util::JavaClass m_io_realm_sync_exception;
+    jni_util::JavaClass m_io_realm_mongodb_app_exception;
     jni_util::JavaClass m_io_realm_sync_log_callback;
     jni_util::JavaClass m_io_realm_sync_error_callback;
 
@@ -102,14 +102,14 @@ public:
         return instance()->m_io_realm_long_pointer_wrapper;
     }
 
-    inline static const jni_util::JavaClass& app_exception()
-    {
-        return instance()->m_io_realm_app_exception;
-    }
-
     inline static const jni_util::JavaClass& sync_exception()
     {
         return instance()->m_io_realm_sync_exception;
+    }
+
+    inline static const jni_util::JavaClass& app_exception_class()
+    {
+        return instance()->m_io_realm_mongodb_app_exception;
     }
 
     inline static const jni_util::JavaClass& sync_log_callback()