+++ packages/jni-swig-stub/src/main/jni/realm_api_helpers.cpp
@@ -219,14 +219,9 @@ realm_t *open_realm_with_scheduler(int64_t config_ptr, jobject dispatchScheduler
 
 void register_login_cb(realm_app_t *app, realm_app_credentials_t *credentials, jobject callback) {
     auto jenv = get_env();
-    // TODO OPTIMIZE Makes multiple lookups of CinteropCallback, but at least only once when
-    //  initializing static variables
-    //  https://github.com/realm/realm-kotlin/issues/460
-    static jmethodID on_success_method = lookup(jenv, "io/realm/internal/interop/CinteropCallback",
-                                                "onSuccess",
-                                                "(Lio/realm/internal/interop/NativePointer;)V");
-    static jmethodID on_error_method = lookup(jenv, "io/realm/internal/interop/CinteropCallback",
-                                              "onError", "(Ljava/lang/Throwable;)V");
+    static JavaClass cinterop_callback(jenv, "io/realm/internal/interop/CinteropCallback");
+    static JavaMethod on_success_method(jenv, cinterop_callback, "onSuccess", "(Lio/realm/internal/interop/NativePointer;)V");
+    static JavaMethod on_error_method(jenv, cinterop_callback, "onError", "(Ljava/lang/Throwable;)V");
 
     realm_app_log_in_with_credentials(
             app,
@@ -241,20 +236,29 @@ void register_login_cb(realm_app_t *app, realm_app_credentials_t *credentials, j
                                          on_error_method,
                                          jenv->ExceptionOccurred());
                 } else if (error) {
-                    // TODO OPTIMIZE Make central global reference table of classes
-                    //  https://github.com/realm/realm-kotlin/issues/460
-                    jclass exception_class = jenv->FindClass("io/realm/mongodb/AppException");
-                    static jmethodID exception_constructor = jenv->GetMethodID(exception_class, "<init>",
-                                                                               "()V");
+                    static JavaMethod app_exception_constructor(jenv,
+                                                                JavaClassGlobalDef::app_exception_class(),
+                                                                "<init>",
+                                                                "(Ljava/lang/String;)V");
 
-                    jobject throwable = jenv->NewObject(exception_class, exception_constructor);
+                    std::stringstream message;
+                    message << error->message << " ["
+                            << "error_category=" << error->error_category << ", "
+                            << "error_code=" << error->error_code << ", "
+                            << "link_to_server_logs=" << error->link_to_server_logs
+                            << "]";
 
+                    jobject throwable = jenv->NewObject(JavaClassGlobalDef::app_exception_class(),
+                                                        app_exception_constructor,
+                                                        to_jstring(jenv, message.str()));
                     jenv->CallVoidMethod(static_cast<jobject>(userdata),
                                          on_error_method,
                                          throwable);
                 } else {
                     // TODO OPTIMIZE Make central global reference table of classes
                     //  https://github.com/realm/realm-kotlin/issues/460
+                    //  Use wrap_pointer of https://github.com/realm/realm-kotlin/pull/493/files#diff-c71ff0791782da90014db6f7000bdeea6c883ca9729f17bd1b11041d1e7711faR450
+                    //  when merged
                     jclass exception_class = jenv->FindClass("io/realm/internal/interop/LongPointerWrapper");
                     static jmethodID exception_constructor = jenv->GetMethodID(exception_class, "<init>", "(JZ)V");
 
@@ -361,19 +365,18 @@ static void pass_jvm_response_to_core(JNIEnv *jenv, jobject j_response, void* re
             j_response, get_headers_method)));
 
     auto stacked_headers = std::vector<std::string>(); // Pins headers to function stack
-    auto response_headers = std::vector<realm_http_header_t>();
-
     for (int i = 0; i < java_headers.size(); i = i + 2) {
         JStringAccessor key = java_headers[i];
         JStringAccessor value = java_headers[i + 1];
         stacked_headers.push_back(std::move(key));
         stacked_headers.push_back(std::move(value));
-
+    }
+    auto response_headers = std::vector<realm_http_header_t>();
+    for (int i = 0; i < java_headers.size(); i = i + 2) {
         // FIXME REFACTOR when C++20 will be available
         realm_http_header header;
         header.name = stacked_headers[i].c_str();
         header.value = stacked_headers[i + 1].c_str();
-
         response_headers.push_back(header);
     }