+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -4,6 +4,21 @@ import io.realm.log.LogLevel
 import io.realm.log.RealmLogger
 
 /**
+ * Runtime identifier. Either 'JVM' or 'Native'
+ */
+expect val RUNTIME: String
+
+/**
+ * Operation system name.
+ */
+expect val OS_NAME: String
+
+/**
+ * Operation system version.
+ */
+expect val OS_VERSION: String
+
+/**
  * Returns the root directory of the platform's App data.
  */
 expect fun appFilesDirectory(): String