+++ packages/library-base/src/darwin/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -7,10 +7,17 @@ import kotlinx.cinterop.alloc
 import kotlinx.cinterop.memScoped
 import kotlinx.cinterop.ptr
 import kotlinx.cinterop.value
+import platform.Foundation.NSProcessInfo
 import platform.posix.pthread_threadid_np
 import kotlin.native.concurrent.ensureNeverFrozen
 import kotlin.native.concurrent.freeze
 import kotlin.native.concurrent.isFrozen
+
+@Suppress("MayBeConst") // Cannot make expect/actual const
+actual val RUNTIME: String = "Native"
+// These causes memory mapping rendering MemoryTests to fail, so only initialize them if actually needed
+actual val OS_NAME: String by lazy { NSProcessInfo.Companion.processInfo().operatingSystemName() }
+actual val OS_VERSION: String by lazy { NSProcessInfo.Companion.processInfo().operatingSystemVersionString }
 
 actual fun createDefaultSystemLogger(tag: String, logLevel: LogLevel): RealmLogger =
     NSLogLogger(tag, logLevel)