+++ packages/library-base/src/jvm/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -1,5 +1,8 @@
 package io.realm.internal.platform
 
+@Suppress("MayBeConst") // Cannot make expect/actual const
+actual val RUNTIME: String = "JVM"
+
 actual fun threadId(): ULong {
     return Thread.currentThread().id.toULong()
 }