+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/internal/AppConfigurationImpl.kt
@@ -23,6 +23,9 @@ import io.realm.internal.interop.NativePointer
 import io.realm.internal.interop.RealmInterop
 import io.realm.internal.interop.sync.MetadataMode
 import io.realm.internal.interop.sync.NetworkTransport
+import io.realm.internal.platform.OS_NAME
+import io.realm.internal.platform.OS_VERSION
+import io.realm.internal.platform.RUNTIME
 import io.realm.internal.platform.freeze
 import io.realm.mongodb.AppConfiguration
 import io.realm.mongodb.AppConfiguration.Companion.DEFAULT_BASE_URL
@@ -45,17 +48,25 @@ internal class AppConfigurationImpl(
         timeoutMs = 5000,
         dispatcher = networkTransportDispatcher,
         logger = object : Logger {
+            // This should ideally be the AppConfigurationImpl.log but iOS ktor client seems to
+            // pass all configuration to another thread which freezes it. Having two logs with the
+            // same configuration is not ideal but would yield the same result when the
+            // configuration and loggers itself are frozen.
+            val log: RealmLog = RealmLog(configuration = logConfig)
             override fun log(message: String) {
-                this@AppConfigurationImpl.log.debug(message)
+                this.log.debug(message)
             }
         }
-    )
+    ).freeze() // Kotlin network client needs to be frozen before passed to the C-API
 
     // Only freeze anything after all properties are setup as this triggers freezing the actual
     // AppConfigurationImpl instance itself
     val nativePointer: NativePointer = RealmInterop.realm_app_config_new(
         appId = appId,
         baseUrl = baseUrl,
-        networkTransport = RealmInterop.realm_network_transport_new(networkTransport)
+        networkTransport = RealmInterop.realm_network_transport_new(networkTransport),
+        platform = "$OS_NAME/$RUNTIME",
+        platformVersion = OS_VERSION,
+        sdkVersion = io.realm.internal.SDK_VERSION
     ).freeze()
 }