+++ test/base/src/androidTest/kotlin/io/realm/test/PlatformInfoTest.kt
@@ -14,14 +14,19 @@
  * limitations under the License.
  */
 
-package io.realm.mongodb
+package io.realm.test
 
-/**
- * A **user** holds the user's metadata and tokens for accessing Realm App functionality.
- *
- * The user is used to configure synchronized realms with [SyncConfiguration.Builder].
- *
- * @see App.login
- * @see SyncConfiguration.Builder
- */
-interface User
+import io.realm.internal.platform.OS_NAME
+import io.realm.internal.platform.OS_VERSION
+import io.realm.internal.platform.RUNTIME
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class PlatformInfoTest {
+    @Test
+    fun platformInfo() {
+        assertEquals("JVM", RUNTIME)
+        assertEquals("Android", OS_NAME)
+        assertEquals(android.os.Build.VERSION.RELEASE, OS_VERSION)
+    }
+}